feat: Add token issuer to request headers
There is a requirement to use advanced routing when accessing the
profile service in order to support shared services between
environments. The routing can use headers to determine where to route
the request.
Add a token issuer header when using the profile client, this allows the
advanced routing to determine where the token was generated and as such
where profile requests should be routed.

TIS21-1580