docs: Add TODO to review auditing implementation
The current AuditingAspect implementation has been flagged as a
potential issue due to the increased accessibility of methods and fields
achieved via reflection. While this is not seen as a security issue in
this use case it should be reviewed for better implementation at a later
date, add a TODO so this can be tracked in Sonar.

TIS21-2049