refactor: replace Spring Security JwtHelper
The Spring Security JwtHelper has been deprecated and removed in later
version of Spring Security. The recommended replacement is
`nimbus-jose-jwt`.

While this is not an immediate issue for TIS services, it's use in the
client is a blocker when trying to get the `profile-client` working with
later version of Spring.
Considering that this migration will eventually be needed anyway, doing
the refactoring now is a light touch to resolve forward-compatibility
issues.

TIS21-6233
TIS21-6308